--- src/test/java/com/example/server/RedisTemplateTest.java
@@ -1,73 +1,0 @@
-package com.example.server;
-
-import static org.assertj.core.api.AssertionsForInterfaceTypes.*;
-
-import java.util.Map;
-import java.util.Set;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.redis.core.HashOperations;
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.SetOperations;
-import org.springframework.data.redis.core.ValueOperations;
-
-@SpringBootTest
-public class RedisTemplateTest {
-
-	@Autowired
-	private RedisTemplate<String, String> redisTemplate;
-
-	@Test
-	void testStrings() {
-		// given
-		ValueOperations<String, String> valueOperations = redisTemplate.opsForValue();
-		String key = "stringKey";
-
-		// when
-		valueOperations.set(key, "hello");
-
-		// then
-		String value = valueOperations.get(key);
-		assertThat(value).isEqualTo("hello");
-	}
-
-	@Test
-	void testSet() {
-		// given
-		SetOperations<String, String> setOperations = redisTemplate.opsForSet();
-		String key = "setKey";
-
-		// when
-		setOperations.add(key, "h", "e", "l", "l", "o");
-
-		// then
-		Set<String> members = setOperations.members(key);
-		Long size = setOperations.size(key);
-
-		assertThat(members).containsOnly("h", "e", "l", "o");
-		assertThat(size).isEqualTo(4);
-	}
-
-	@Test
-	void testHash() {
-		// given
-		HashOperations<String, Object, Object> hashOperations = redisTemplate.opsForHash();
-		String key = "hashKey";
-
-		// when
-		hashOperations.put(key, "hello", "world");
-
-		// then
-		Object value = hashOperations.get(key, "hello");
-		assertThat(value).isEqualTo("world");
-
-		Map<Object, Object> entries = hashOperations.entries(key);
-		assertThat(entries.keySet()).containsExactly("hello");
-		assertThat(entries.values()).containsExactly("world");
-
-		Long size = hashOperations.size(key);
-		assertThat(size).isEqualTo(entries.size());
-	}
-}
